# frontend/config/swagger.yaml
swagger: '2.0'
info:
  description: 'This is the theme parks API spec.'
  version: '1.0.0'
  title: 'Themepark API'
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'com.themeparks.api'
...
schemes:
- 'https'
- 'http'
paths:
...
  /reservations:
    get:
      tags:
        - reservation
      operationId: getReservations
      parameters:
      - name: userId
        in: query
        description: user account
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservations'
    post:
      tags:
        - reservation
      description: create new reservations
      operationId: createReservation
      parameters:
      - name: body
        in: body
        description: reservation
        schema:
          $ref: '#/definitions/Reservations'
        required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservations'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      x-codegen-request-body-name: body
    delete:
      tags:
        - reservation
      description: delete reservation
      operationId: deleteReservation
      parameters:
        - name: reservationId
          in: query
          description: id of reservation to delete
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservations'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /reservations/confirm:
    post:
      tags:
        - reservation
      description: confirm reservation
      operationId: confirmReservation
      parameters:
      - name: reservationId
        in: query
        description: id of reservation to confirm
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservation'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /queue:
    get:
      tags:
        - queue
      operationId: getCurrentQueue
      parameters:
      - name: userId
        in: query
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Queue'
  /queue/join:
    post:
      tags:
        - queue
      description: join a queue
      operationId: joinNewQueue
      parameters:
      - name: body
        in: body
        description: queue info
        schema:
          $ref: '#/definitions/JoinQueueRequest'
        required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Queue'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /queue/quit:
    delete:
      tags:
        - queue
      description: quit from a queue
      operationId: quitQueue
      parameters:
      - name: user
        in: query
        description: user ID who wants to quit his/her queue.
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Queue'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      x-codegen-request-body-name: body
...
  /account/logout:
    post:
      description: log out of the system
      operationId: logout
      tags:
        - login
      responses:
        '200':
          description: Success
          schema:
            type: object
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      parameters: []
      produces:
        - application/json
    x-swagger-router-controller: api
  /account/login:
    post:
      tags:
        - login
      description: log into the system
      operationId: login
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/LoginResult'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      x-codegen-request-body-name: body
      parameters:
        - description: login params
          required: true
          name: body
          in: body
          schema:
            $ref: '#/definitions/LoginParams'
      consumes:
        - application/json
      produces:
        - application/json
    x-swagger-router-controller: api
...
definitions:
...
  Reservations:
    type: array
    items:
      $ref: '#/definitions/Reservation'
  Reservation:
    type: object
    properties:
      id:
        type: string
      attraction:
        type: string
      user:
        type: string
      schedule:
        type: string
      startTime:
        type: string
        format: date-time
      endTime:
        type: string
        format: date-time
      location:
        type: string
      status:
        type: string
        enum: [waiting, passed, cancelled, confirmed, checking-in]
  Queue:
    type: object
    properties:
      id:
        type: string
      numberInQueue:
        type: integer
      nextNumberToCall:
        type: integer
      attraction:
        type: string
      user:
        type: string
      location:
        type: string
      eta:
        type: string
        format: date-time
      status:
        type: string
        enum: [waiting, passed, called, confirmed]
  JoinQueueRequest:
    type: object
    properties:
      attraction:
        type: string
      users:
        type: array
        items:
          type: string
  ErrorResponse:
    required:
      - errorCode
    type: object
    properties:
      errorCode:
        type: string
        description: error code
      errorMessage:
        type: string
        description: error message
      success:
        type: boolean
        description: success or not
...
  LoginResult:
    type: object
    properties:
      status:
        type: string
      type:
        type: string
      currentAuthority:
        type: string
  LoginParams:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
      autoLogin:
        type: boolean
      type:
        type: string